[change] Changes updated #576
Fixes #576
