+++ openwisp_monitoring/device/tests/test_apps.py
@@ -1,6 +1,9 @@
 from django.test import TestCase
 
+from openwisp_monitoring.device.models import CHECK_CLASSES, Check
+from openwisp_monitoring.device.signals import health_status_changed
 from openwisp_utils.admin_theme.dashboard import DASHBOARD_CHARTS, DASHBOARD_TEMPLATES
+from openwisp_utils.tests import catch_signal
 
 
 class TestApps(TestCase):
@@ -37,3 +40,54 @@ class TestApps(TestCase):
                 'monitoring/js/device-map.js',
             ),
         )
+
+
+class TestCheckSignals(TestCase):
+    def setUp(self):
+        self.device = self._create_device(organization=self._create_org())
+        self.ping_check = Check.objects.create(
+            name='Ping Check',
+            check_type=CHECK_CLASSES[0][0],
+            content_object=self.device,
+            params={},
+        )
+        self.device.monitoring.update_status('ok')
+
+    def test_check_signals(self):
+        with self.subTest('Test disabling a critical check'):
+            self.assertEqual(self.device.monitoring.status, 'ok')
+            with catch_signal(health_status_changed) as handler:
+                self.ping_check.is_active = False
+                self.ping_check.save()
+            self.assertEqual(handler.call_count, 1)
+            call_args = handler.call_args[0]
+            self.assertEqual(call_args[0], self.device.monitoring)
+            self.assertEqual(call_args[1], 'unknown')
+            self.device.refresh_from_db()
+            self.assertEqual(self.device.monitoring.status, 'unknown')
+            self.device.monitoring.update_status('ok')
+
+        with self.subTest('Test saving an active critical check'):
+            self.assertEqual(self.device.monitoring.status, 'ok')
+            self.ping_check.is_active = True
+            self.ping_check.save()
+            self.device.refresh_from_db()
+            self.assertEqual(self.device.monitoring.status, 'ok')
+
+        with self.subTest('Test saving a non-critical check'):
+            self.assertEqual(self.device.monitoring.status, 'ok')
+            non_critical_check = Check.objects.create(
+                name='Configuration Applied',
+                check_type=CHECK_CLASSES[1][0],
+                content_object=self.device,
+                params={},
+            )
+            non_critical_check.save()
+            self.device.refresh_from_db()
+            self.assertEqual(self.device.monitoring.status, 'ok')
+
+        with self.subTest('Test deleting a critical check'):
+            self.device.monitoring.update_status('ok')
+            self.ping_check.delete()
+            self.device.refresh_from_db()
+            self.assertEqual(self.device.monitoring.status, 'unknown')
